refactor: remove @Deprecated methods (#3184)
* up

* up

* @monperrus at work 🗣

* @monperrus at work 🤜

* @monperrus at work 🤢

* @monperrus at work 💻

* @monperrus at work 🏇
