--- src/main/java/.cvsignore
@@ -1,1 +1,0 @@
-com*
