+++ src/main/java/spoon/reflect/cu/CompilationUnit.java
@@ -11,38 +11,6 @@ import spoon.reflect.declaration.CtCompilationUnit;
  * Defines a compilation unit. In Java, a compilation unit can contain only one
  * public type declaration and other secondary types declarations (not public).
  */
+@Deprecated
 public interface CompilationUnit extends CtCompilationUnit {
-
-
-	/**
-	 * Helper method to get the begin index of the line that corresponds to the
-	 * given index.
-	 *
-	 * @param index
-	 * 		an arbitrary index in the source code
-	 * @return the index where the line starts
-	 */
-	@Deprecated
-	int beginOfLineIndex(int index);
-
-	/**
-	 * Helper method to get the begin index of the line that corresponds to the
-	 * next line of the given index.
-	 *
-	 * @param index
-	 * 		an arbitrary index in the source code
-	 * @return the index where the next line starts
-	 */
-	@Deprecated
-	int nextLineIndex(int index);
-
-	/**
-	 * Gets the number of tabulations for a given line.
-	 *
-	 * @param index
-	 * 		the index where the line starts in the source code
-	 * @return the number of tabs for this line
-	 */
-	@Deprecated
-	int getTabCount(int index);
 }
